requirements.txt
asgiref==3.3.4
attrs==21.2.0
Django==3.2.3
djangorestframework==3.12.4
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
